requirements.txt
streamlit
numpy
pillow
scikit-learn
anyio  # if needed